refactor: Migrate face overlay logic to cl_media_viewer, update extended_image, and add image scale change callback.

+++ cl_media_viewer/example/lib/main.dart
@@ -351,8 +351,8 @@ class _ImageViewerScreenState extends State<ImageViewerScreen> {
     return InteractiveImageViewer(
       imageData: InteractiveImageData(
         uri: Uri.parse('asset:${imageData.assetPath}'),
-        width: imageData.width,
-        height: imageData.height,
+        width: imageData.width.toDouble(),
+        height: imageData.height.toDouble(),
         faces: imageData.faces.map((faceData) {
           return InteractiveFace(
             data: faceData,
